refactor explicit canonical model compiler
